Update ProvTabs file structure and pull TabPanel into its own file
