+++ context/app/static/js/components/Detail/ProvTabs/index.js
@@ -1,0 +1,3 @@
+import ProvTabs from './ProvTabs';
+
+export default ProvTabs;
