+++ context/app/static/js/components/Detail/TabPanel/index.js
@@ -1,0 +1,3 @@
+import TabPanel from './TabPanel';
+
+export default TabPanel;
